# secp256k1: Use mod n scalar in sig serialization.