**This requires PR #2085**.

This modifies the function to serialize signatures to use the new `ModNScalar` type instead of big ints and takes the opportunity to clean up the relevant code so that it is more readable and avoid unnecessary extra allocations, improves the comments to reference the DER specification and include the specific details of the format inline, and adds a signature serialization benchmark.

This is work towards eventually using the new more efficient mod n scalar throughout.

While the primary focus of this change is not optimization since signature serialization happens infrequently enough that it is not in a performance hot path, as can be seen from the benchmark below, it does have the effect of making the serialization faster and significantly reducing the number of allocations.

The following benchmark shows a before and after comparison of serializing a typical signature:

```
benchmark               old ns/op    new ns/op    delta
 ---------------------------------------------------------
BenchmarkSigSerialize   373          316          -15.28%

benchmark               old allocs   new allocs   delta
---------------------------------------------------------
BenchmarkSigSerialize   5            3            -40.00%

benchmark               old bytes    new bytes    delta
---------------------------------------------------------
BenchmarkSigSerialize   256          144          -43.75%
```